Replace unlicenced bc module with GPL-3 code
Lots of code from this commit was taken from
https://github.com/aboood40091/BNTX-Editor, which is fully licensed
under GPL-3, allowing the whole project to be licensed as GPL-3.0
allowing it to be submitted as a proper blender extension, as well as
making it fully legal to be modified, which I should have done from the
beginning

The new module for BCn textures is much slower than before, and
my implementation of it doesnt mix too well with the existing code. I'm
hoping to both issues over time, but currently it works exactly like it
did before. Just slower